move contains method to support folder of viewManager
